// src/generated/models/AnnotationsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Annotationsobjecttypecode = {
  1: 'Account',
  4201: 'Appointment',
  4407: 'BulkImport',
  4003: 'Calendar',
  4400: 'Campaign',
  4402: 'CampaignActivity',
  4401: 'CampaignResponse',
  112: 'Case',
  4206: 'CaseResolution',
  4215: 'Commitment',
  123: 'Competitor',
  2: 'Contact',
  1010: 'Contract',
  1011: 'ContractLine',
  4202: 'Email',
  4000: 'Facility_Equipment',
  4204: 'Fax',
  1090: 'Invoice',
  4: 'Lead',
  4207: 'Letter',
  4300: 'MarketingList',
  3: 'Opportunity',
  4208: 'OpportunityClose',
  1088: 'Order',
  4209: 'OrderClose',
  4210: 'PhoneCall',
  1024: 'Product',
  1084: 'Quote',
  4211: 'QuoteClose',
  4006: 'ResourceSpecification',
  4001: 'Service',
  4214: 'ServiceActivity',
  4212: 'Task',
  8181: 'RoutingRule',
  8199: 'RoutingRuleItem'
} as const;
export type Annotationsobjecttypecode = keyof typeof Annotationsobjecttypecode;

export interface AnnotationsBase {
  annotationid: string;
  documentbody?: string;
  filename?: string;
  importsequencenumber?: number;
  isdocument: boolean;
  langid?: string;
  mimetype?: string;
  notetext?: string;
  "ObjectId@odata.bind"?: string;
  objectidtypecode?: string;
  objecttypecode?: Annotationsobjecttypecode;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  stepid?: string;
  subject: string;
}

export interface Annotations extends AnnotationsBase {
  createdbyname?: string;
  createdbyyominame?: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame?: string;
  dummyfilename?: string;
  dummyregarding?: string;
  filepointer?: string;
  filesize?: number;
  iscompressed?: boolean;
  iscompressedname?: string;
  isdocumentname?: string;
  isprivate?: boolean;
  isprivatename?: string;
  modifiedbyname?: string;
  modifiedbyyominame?: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame?: string;
  objecttypecodename?: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname?: string;
  prefix?: string;
  storagepointer?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  objectid?: object;
  _objectid_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}